Fielder Covering Bases Improvement
Fielders now throw to the base instead of the current position of the player to be covering. Players now correctly cover other bases. Temporarily disabled fielders' collisions with each other. The fielders' collision should be rethought to allow for moving around each other. The raycast system is bad. This can be done with a different collision detection setting on the rigibody or possibly learning how to use NavMesh. Also players should be able to move instead of throwing a very small distance. I think I should change that.